First useable version
The tools can now be used for basic comparison, I will probably not
make many improvements from now on.  It would be useful to compare
wind fields, which currently does not work for some reason (problem
when unskewing in wrf_user_getvar().

Otherwise, it seems to be working as it should.